% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_survey.R
\docType{data}
\name{stack_survey}
\alias{stack_survey}
\title{Results of the Stack Overflow 2016 Developer Survey}
\format{One row for each respondent (56,030), with 66 columns: one column for each
survey question (along with some columns added through processing). The format and
text for each question is described in the \code{\link{stack_schema}} dataset.}
\source{
\url{http://stackoverflow.com/research}
}
\usage{
stack_survey
}
\description{
Developer Survey results as a data frame (specifically a tbl_df). While this is
almost identical to the CSV available from \url{http://stackoverflow.com/research},
this takes the additional step of turning each range (e.g. years of experience being
"1 - 5 years") and each scale (e.g. "On a scale of 1 to 5, how important...") into an
ordered factor, which allows easier
visualization and sorting.
}
\keyword{datasets}

